New Physical Memory for x86-ISA Model
This books define an abstract STOBJ called @('mem') that exports an
accessor function @('read-mem') and an updater function @('write-mem') --
just like @('bigmem'), library @('mem') models a byte memory of arbitrary
size.  The corresponding concrete STOBJ for the @('bigmem2') @('mem') is
composed of two fields: an array of bytes addressed by natural numbers
smaller than a threshold and a hashtable that writes-and-reads bytes on
demand.</p>

<p>An obvious application of @('bigmem2') is to model a byte-wide
memory; @('mem2') can be used as a child STOBJ to define a field
representing a byte memory in a parent stobj that models some
machine's state.  See @('projects/x86isa/machine/state.lisp') for one
such@('bigmem') example.")